bug fixes and base functionality to improve to the crawler ecologylab/semantics/actions   EvaluateRankWeight, ParseDocumentSemanticAction, SemanticActionNamedArguments -- added attributes and minor bugfix ecologylab/semantics/connectors   InfoCollectorBase -- set the session to either being heterogeneous or homogeneous as per the current seed set ecologylab/semantics/documentparsers   HTMLDOMImageTextParser, HTMLFragmentDOMParser -- make code navigable towards JTidy ecologylab/semantics/html/documentstructure   ImageCollectionPage -- todo   SemanticInLinks -- reduce emphasis on anchors from same pages. ecologylab/semantics/model/text/utils   Filter -- add 'doublclick' as an ad pattern ecologylab/semantics/seeding   DocumentState, SearchState, Seed -- provide appropriate response to isHomogeneous seed. repository   agentsEnginesStylesSites.xml -- sorted search_engines, made (but not implemented) site specific search engins repository/repositorySources   wikipedia.xml -- only pick top 'n' documents from a page to crawl into.

+++ ecologylab/semantics/actions/EvaluateRankWeight.java
@@ -51,9 +51,9 @@ class EvaluateRankWeight<IC extends InfoCollector, SAH extends SemanticActionHan
 		float result = ((float) size - index) / size;
 
 		double e = Math.E;
-		double val = 1 / (1 + Math.pow(e, CURVE_AMOUNT * (TRANSFER_FUNC_CENTER - result)));
+		float val = (float)( 1 / (1 + Math.pow(e, CURVE_AMOUNT * (TRANSFER_FUNC_CENTER - result))));
 
-		return Double.valueOf(val);
+		return Float.valueOf(val);
 	}
 
 }
